refactor(cli): move OpenAI Evals import into src/importers/ module
Splits ~470 lines of OpenAI-specific conversion logic out of
src/commands/import.ts into a dedicated src/importers/openaiEvals/
module so the CLI command stays a thin shim and the conversion is
reusable from non-CLI callers (web UI, programmatic API).

Structure:
- src/importers/openaiEvals/types.ts: OpenAIEvalsJsonlRow, OpenAISample
  (new — centralizes the sample shape used by 4 helpers),
  OpenAIEvalsImportResult (new — types the conversion output for the
  discriminated union)
- src/importers/openaiEvals/validation.ts: isRecord, isOpenAIEvalsJsonlRow
- src/importers/openaiEvals/convert.ts: conversion helpers and the
  single-pass convertOpenAIEvalsJsonl entry point
- src/importers/openaiEvals/index.ts: public surface
- src/importers/types.ts: ParsedImportFile as a discriminated union
  on `source` (replaces the previous `evalData: any` with typed
  OpenAI branch + `unknown` Promptfoo branch)
- src/importers/parse.ts: parseImportFile router with narrowed try
  so downstream conversion errors no longer surface as JSON syntax
  errors

Behavior-preserving improvements bundled in:
- Single-pass conversion loop replaces four sequential passes over
  rows (eliminates O(N*M) run/prompt filters and double-canonicalize)
- Convention-compliant error stringification in the top-level catch
  (matches sibling commands in src/commands/auth.ts, logs.ts, eval.ts)
- logger.debug surfaces for previously-silent drops: malformed
  token_usage, unrecognized sample.error shape, non-array outputs,
  per-line JSONL parse failures
- Unrecognized sample.error shapes now JSON.stringify'd into the
  result error instead of silently discarded
- Hoisted string constants (provider prefix, import format, source
  discriminator values)
- Extracted createOpenAITestCase + getOpenAIFailureReason helpers
- Comment polish: outputItem duplicate marked as intentional
  forward-compat stash

Test coverage additions in test/commands/import.test.ts:
- Canonicalization aligns reordered-key items into one testIdx
  (the headline alignment feature was previously untested)
- Mixed valid/invalid JSONL lines surface the parse error and abort
  the import with exit code 1
- Unsafe characters in run_id are sanitized in the generated eval id
- ResultFailureReason enum replaces magic-number comparisons
